src/backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0

celery==5.6.2
redis==5.3.1

psycopg2-binary==2.9.10
python-decouple==3.8

pillow==11.3.0
qrcode==8.2

gunicorn
